Complete guessing game
Adds guessing game. A guessing game is created with an answer and
an isSolved boolean defaulting to false. Guess function exists to
allow a user to guess at the answer. Guess will trigger responses
of low or high from the guess; these functions use the highGuess
and lowGuess guessingGame functions. A correct guess will congratulate
the user and set the isSolved boolean for the game to true. All
tests are currently passing.